Fix user model to properly unselect token fields by default
This is very important as it appears that mongoose will work without explicitly setting the type of a nested object. However, the select=false settings in CharacterModel will not be respected then. It is necessary to set { type: [Character] } for the alts array to apply the select settings to a list of objects.